Moved class ActiveEntityModifier to another package
